pyproject.toml
[project]
name = "mdsgene"
version = "0.1.0"
authors = [
    {name = "Gleb Pankov"},
    {name = "Harut Madoev"}
]
description = "MDSGene Application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.10.0",
    "fastapi>=0.115.12",
    "google-cloud-aiplatform>=1.87.0",
    "google-genai>=1.10.0",
    "ipywidgets>=8.1.5",
    "langchain-community>=0.3.20",
    "langchain-ollama>=0.3.0",
    "langgraph>=0.4.3",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "paper-qa>=5.11.1",
    "pypdf>=5.4.0",
    "ruff>=0.11.8",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "jupyterlab>=4.3.6",
]
